[MEASURE SAVER] Unit tests: moved general stuff to separated context
